[LOCAL][UPDATE] Handle larger exit codes

+++ convert.c
@@ -5,11 +5,11 @@
  * @s: input string
  * Return: converted int
  */
-int _atoi(char *s)
+long _atoi(char *s)
 {
-	int loop, digitCount = 0, neg = 1, start = 0, multiple = 1;
+	long loop, digitCount = 0, neg = 1, start = 0, multiple = 1;
 
-	int temp, number = 0;
+	long temp, number = 0;
 
 	for (loop = 0; s[loop] != '\0'; loop++)
 	{
